Remove num_workers option from validation DataLoader
Setting the `num_workers` option to the validation DataLoader does not have any benefit (at least for pytorch 1.8+).
If this is set it doesn't allow to specify number of CPU threads in case one wants to limit the CPU usage
i.e. `torch.set_num_threads(8)` will break the execution.